# docs/openapi-v1.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lar Finance Private API",
    "version": "1.0.0"
  },
  "servers": [{"url": "/api/v1"}],
  "security": [{"opaqueBearer": []}],
  "paths": {
...
    "/cards/{id}/": {
      "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}],
      "get": {
        "operationId": "getCardDetail",
        "security": [{"opaqueBearer": []}],
        "responses": {
          "200": {"description": "Card details and selected invoice", "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestId"}}},
          "401": {"$ref": "#/components/responses/Unauthorized"}
        }
      },
      "put": {
        "operationId": "updateCardPut",
        "security": [{"opaqueBearer": []}],
        "responses": {
          "200": {"description": "Updated card", "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestId"}}},
          "401": {"$ref": "#/components/responses/Unauthorized"}
        }
      },
      "patch": {
        "operationId": "updateCardPatch",
        "security": [{"opaqueBearer": []}],
        "responses": {
          "200": {"description": "Updated card", "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestId"}}},
          "401": {"$ref": "#/components/responses/Unauthorized"}
        }
      },
      "delete": {
        "operationId": "deleteCard",
        "security": [{"opaqueBearer": []}],
        "responses": {
          "200": {"description": "Archived card", "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestId"}}},
          "401": {"$ref": "#/components/responses/Unauthorized"}
        }
      }
    },
...
  },
  "components": {
    "headers": {
      "RequestId": {
        "description": "UUID used to correlate the request with its access log.",
        "schema": {"type": "string", "format": "uuid"}
      }
    },
    "securitySchemes": {
      "opaqueBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Opaque access token issued for one revocable device session."
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Missing, invalid, expired, or revoked device credentials",
        "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestId"}},
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}
      },
...
    },
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "required": ["error", "request_id"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message", "fields"],
            "properties": {
              "code": {"type": "string"},
              "message": {"type": "string"},
              "fields": {"type": ["object", "array", "null"]}
            }
          },
          "request_id": {"type": "string", "format": "uuid"}
        }
      },
...
    }
  }
}